+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * #72 Fix running javamaker with spaces in project path
 * Load description.xml editor even when locale is missing
 * Remove crashreport service (Service no longer exists)
+* Fix build with empty `build` directory
 
 ## 3.0.0
 * Python support for extension development